requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
defusedxml==0.8.0rc2
dj-rest-auth==7.0.1
Django>=4.2.17, <5.0.0
django-allauth==65.3.1
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.0rc2
Markdown==3.7
mutagen==1.47.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn==1.6.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
rsa==4.9
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
zipp==3.21.0
pytz~=2025.1